+++ modules/example/controllers/manage_hobby.php
@@ -84,6 +84,14 @@ class Manage_Hobby extends CMS_Priv_Strict_Controller {
         // displayed columns on add operation
         $crud->add_fields('name');
 
+        //////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+        // HINT: Put Tabs (if needed)
+        // usage:
+        //     $crud->set_tabs(array(
+        //        'First Tab Caption'  => $how_many_field_on_first_tab,
+        //        'Second Tab Caption' => $how_many_field_on_second_tab,
+        //     ));
+        //////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 
         // caption of each columns
@@ -180,26 +188,35 @@ class Manage_Hobby extends CMS_Priv_Strict_Controller {
         if(!$crud->unset_delete){
             $id_list = json_decode($this->input->post('data'));
             foreach($id_list as $id){
-                $this->db->delete($this->cms_complete_table_name('hobby'),array('hobby_id'=>$id));
+                if($this->_before_delete($id)){
+                    $this->db->delete($this->cms_complete_table_name('hobby'),array('hobby_id'=>$id));
+                    $this->_after_delete($id);
+                }
             }
         }
     }
 
     public function _before_insert($post_array){
+        $post_array = $this->_before_insert_or_update($post_array);
+        // HINT : Put your code here
         return $post_array;
     }
 
     public function _after_insert($post_array, $primary_key){
         $success = $this->_after_insert_or_update($post_array, $primary_key);
+        // HINT : Put your code here
         return $success;
     }
 
     public function _before_update($post_array, $primary_key){
+        $post_array = $this->_before_insert_or_update($post_array, $primary_key);
+        // HINT : Put your code here
         return $post_array;
     }
 
     public function _after_update($post_array, $primary_key){
         $success = $this->_after_insert_or_update($post_array, $primary_key);
+        // HINT : Put your code here
         return $success;
     }
 
@@ -215,6 +232,10 @@ class Manage_Hobby extends CMS_Priv_Strict_Controller {
     public function _after_insert_or_update($post_array, $primary_key){
 
         return TRUE;
+    }
+
+    public function _before_insert_or_update($post_array, $primary_key=NULL){
+        return $post_array;
     }
